# openapi/spec.yml
components:
  schemas:
...
    accountView:
      properties:
        account_address:
          type: string
        account_iban:
          type: string
        account_name:
          type: string
        bank_address:
          type: string
        bank_name:
          type: string
        bank_swift:
          type: string
        id:
          format: uuid
          type: string
        label:
          type: string
      type: object
...
    categoryView:
      properties:
        category:
          $ref: '#/components/schemas/categoryView'
        category_id:
          format: uuid
          type: string
        id:
          format: uuid
          type: string
        items:
          items:
            $ref: '#/components/schemas/itemView'
          type: array
        name:
          type: string
      type: object
...
    companyView:
      properties:
        address:
          type: string
        contact:
          type: string
        crn:
          type: string
        email:
          format: email
          type: string
        id:
          format: uuid
          type: string
        mobile:
          type: string
        name:
          type: string
        trn:
          type: string
      type: object
...
    currencyView:
      properties:
        id:
          type: string
        name:
          type: string
        symbol:
          type: string
      type: object
...
    customerView:
      properties:
        address:
          type: string
        contact:
          type: string
        currency:
          $ref: '#/components/schemas/currencyView'
        currency_id:
          type: string
        id:
          format: uuid
          type: string
        name:
          type: string
        payment_term:
          type: integer
        sales:
          items:
            $ref: '#/components/schemas/saleView'
          type: array
      type: object
...
    itemCreate:
      properties:
        category_id:
          format: uuid
          type: string
        name:
          type: string
      required:
      - name
      type: object
    itemPartial:
      properties:
        category_id:
          format: uuid
          type: string
        name:
          type: string
      type: object
...
    itemUpdate:
      properties:
        category_id:
          format: uuid
          type: string
        name:
          type: string
      required:
      - name
      type: object
    itemView:
      properties:
        category:
          $ref: '#/components/schemas/categoryView'
        category_id:
          format: uuid
          type: string
        id:
          format: uuid
          type: string
        name:
          type: string
      type: object
...
    saleCreate:
      properties:
        account_id:
          format: uuid
          type: string
        company_id:
          format: uuid
          type: string
        confirmed:
          type: boolean
        currency_id:
          type: string
        customer_id:
          format: uuid
          type: string
        date:
          format: date
          type: string
        due_date:
          format: date
          type: string
        place:
          type: string
        reference:
          type: string
        reference_date:
          format: date
          type: string
      required:
      - company_id
      - account_id
      - customer_id
      - currency_id
      - date
      - confirmed
      type: object
...
    saleItemCreate:
      properties:
        description:
          type: string
        item_id:
          format: uuid
          type: string
        price:
          type: number
        quantity:
          type: integer
        sale_id:
          format: uuid
          type: string
      required:
      - sale_id
      - item_id
      - quantity
      - price
      type: object
    saleItemPartial:
      properties:
        description:
          type: string
        item_id:
          format: uuid
          type: string
        price:
          type: number
        quantity:
          type: integer
        sale_id:
          format: uuid
          type: string
      type: object
...
    saleItemUpdate:
      properties:
        description:
          type: string
        item_id:
          format: uuid
          type: string
        price:
          type: number
        quantity:
          type: integer
        sale_id:
          format: uuid
          type: string
      required:
      - sale_id
      - item_id
      - quantity
      - price
      type: object
    saleItemView:
      properties:
        description:
          type: string
        id:
          format: uuid
          type: string
        item:
          $ref: '#/components/schemas/itemView'
        item_id:
          format: uuid
          type: string
        price:
          type: number
        quantity:
          type: integer
        sale:
          $ref: '#/components/schemas/saleView'
        sale_id:
          format: uuid
          type: string
      type: object
    salePartial:
      properties:
        account_id:
          format: uuid
          type: string
        company_id:
          format: uuid
          type: string
        confirmed:
          type: boolean
        currency_id:
          type: string
        customer_id:
          format: uuid
          type: string
        due_date:
          format: date
          type: string
        place:
          type: string
        reference:
          type: string
        reference_date:
          format: date
          type: string
      type: object
...
    saleUpdate:
      properties:
        account_id:
          format: uuid
          type: string
        company_id:
          format: uuid
          type: string
        confirmed:
          type: boolean
        currency_id:
          type: string
        customer_id:
          format: uuid
          type: string
        due_date:
          format: date
          type: string
        place:
          type: string
        reference:
          type: string
        reference_date:
          format: date
          type: string
      required:
      - company_id
      - account_id
      - customer_id
      - currency_id
      - confirmed
      type: object
    saleView:
      properties:
        account:
          $ref: '#/components/schemas/accountView'
        account_id:
          format: uuid
          type: string
        company:
          $ref: '#/components/schemas/companyView'
        company_id:
          format: uuid
          type: string
        confirmed:
          type: boolean
        currency:
          $ref: '#/components/schemas/currencyView'
        currency_id:
          type: string
        customer:
          $ref: '#/components/schemas/customerView'
        customer_id:
          format: uuid
          type: string
        date:
          format: date
          type: string
        due_date:
          format: date
          type: string
        id:
          format: uuid
          type: string
        items:
          items:
            $ref: '#/components/schemas/saleItemView'
          type: array
        number:
          type: integer
        place:
          type: string
        reference:
          type: string
        reference_date:
          format: date
          type: string
        total:
          type: number
      type: object
...
  securitySchemes:
    bearer:
      scheme: bearer
      type: http
    cookieAccessToken:
      in: cookie
      name: x-access-token
      type: apiKey
    cookieRefreshToken:
      in: cookie
      name: x-refresh-token
      type: apiKey
    key:
      in: header
      name: X-API-Key
      type: apiKey
    oauth:
      flows:
        authorizationCode:
          authorizationUrl: https://example.com/oauth/authorize
          scopes:
            admin: Grants access to admin operations
            read: Grants read access
            write: Grants write access
          tokenUrl: https://example.com/oauth/token
      type: oauth2
info:
  description: Invoicing API Reference
  title: Invoicing API
  version: '1.0'
openapi: 3.0.3
paths:
...
  /api/v1/items:
    get:
      description: get all item entities
      operationId: getItems
      parameters:
      - description: Filter Items by name
        in: query
        name: name
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/itemView'
                type: array
          description: successful operation
      summary: get all item entities
      tags:
      - item
    post:
      description: Add a new item entity
      operationId: addItem
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/itemCreate'
        description: Create a new item entity
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/itemView'
          description: successful operation
        '405':
          description: invalid input
      summary: Add a new item entity
      tags:
      - item
  /api/v1/items/{id}:
    delete:
      description: Delete an existing item entity
      operationId: deleteItem
      parameters:
      - description: item id to delete
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/itemView'
          description: successful operation
        '400':
          description: Invalid item id
        '404':
          description: item not found
      summary: Delete an existing item entity
      tags:
      - item
    get:
      description: Get an existing item entity
      operationId: getItem
      parameters:
      - description: item id to get
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/itemView'
          description: successful operation
        '400':
          description: Invalid item id
        '404':
          description: item not found
      summary: Get an existing item entity
      tags:
      - item
    patch:
      description: Modify an existing item entity (partial update, only provided properties
        will be updated)
      operationId: modifyItem
      parameters:
      - description: item id to partially update
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/itemPartial'
        description: Modify an existing item entity
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/itemView'
          description: successful operation
        '400':
          description: Invalid ID
        '404':
          description: item not found
        '405':
          description: Invalid input
      summary: Modify an existing item entity
      tags:
      - item
    put:
      description: Update an existing item entity (all required properties must be
        supplied)
      operationId: updateItem
      parameters:
      - description: item id to update
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/itemUpdate'
        description: Update an existing item entity
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/itemView'
          description: successful operation
        '400':
          description: Invalid ID
        '404':
          description: item not found
        '405':
          description: Invalid input
      summary: Update an existing item entity
      tags:
      - item
...
  /api/v1/saleItems:
    get:
      description: get all saleItem entities
      operationId: getSaleitems
      parameters:
      - description: Filter Saleitems by item.name
        in: query
        name: item.name
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/saleItemView'
                type: array
          description: successful operation
      summary: get all saleItem entities
      tags:
      - saleItem
    post:
      description: Add a new saleItem entity
      operationId: addSaleitem
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/saleItemCreate'
        description: Create a new saleItem entity
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/saleItemView'
          description: successful operation
        '405':
          description: invalid input
      summary: Add a new saleItem entity
      tags:
      - saleItem
  /api/v1/saleItems/{id}:
    delete:
      description: Delete an existing saleItem entity
      operationId: deleteSaleitem
      parameters:
      - description: saleItem id to delete
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/saleItemView'
          description: successful operation
        '400':
          description: Invalid saleItem id
        '404':
          description: saleItem not found
      summary: Delete an existing saleItem entity
      tags:
      - saleItem
    get:
      description: Get an existing saleItem entity
      operationId: getSaleitem
      parameters:
      - description: saleItem id to get
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/saleItemView'
          description: successful operation
        '400':
          description: Invalid saleItem id
        '404':
          description: saleItem not found
      summary: Get an existing saleItem entity
      tags:
      - saleItem
    patch:
      description: Modify an existing saleItem entity (partial update, only provided
        properties will be updated)
      operationId: modifySaleitem
      parameters:
      - description: saleItem id to partially update
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/saleItemPartial'
        description: Modify an existing saleItem entity
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/saleItemView'
          description: successful operation
        '400':
          description: Invalid ID
        '404':
          description: saleItem not found
        '405':
          description: Invalid input
      summary: Modify an existing saleItem entity
      tags:
      - saleItem
    put:
      description: Update an existing saleItem entity (all required properties must
        be supplied)
      operationId: updateSaleitem
      parameters:
      - description: saleItem id to update
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/saleItemUpdate'
        description: Update an existing saleItem entity
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/saleItemView'
          description: successful operation
        '400':
          description: Invalid ID
        '404':
          description: saleItem not found
        '405':
          description: Invalid input
      summary: Update an existing saleItem entity
      tags:
      - saleItem
  /api/v1/sales:
    get:
      description: get all sale entities
      operationId: getSales
      parameters:
      - description: Filter Sales by date
        in: query
        name: date
        schema:
          type: string
      - description: Filter Sales by number
        in: query
        name: number
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/saleView'
                type: array
          description: successful operation
      summary: get all sale entities
      tags:
      - sale
    post:
      description: Add a new sale entity
      operationId: addSale
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/saleCreate'
        description: Create a new sale entity
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/saleView'
          description: successful operation
        '405':
          description: invalid input
      summary: Add a new sale entity
      tags:
      - sale
  /api/v1/sales/{id}:
    delete:
      description: Delete an existing sale entity
      operationId: deleteSale
      parameters:
      - description: sale id to delete
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/saleView'
          description: successful operation
        '400':
          description: Invalid sale id
        '404':
          description: sale not found
      summary: Delete an existing sale entity
      tags:
      - sale
    get:
      description: Get an existing sale entity
      operationId: getSale
      parameters:
      - description: sale id to get
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/saleView'
          description: successful operation
        '400':
          description: Invalid sale id
        '404':
          description: sale not found
      summary: Get an existing sale entity
      tags:
      - sale
    patch:
      description: Modify an existing sale entity (partial update, only provided properties
        will be updated)
      operationId: modifySale
      parameters:
      - description: sale id to partially update
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/salePartial'
        description: Modify an existing sale entity
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/saleView'
          description: successful operation
        '400':
          description: Invalid ID
        '404':
          description: sale not found
        '405':
          description: Invalid input
      summary: Modify an existing sale entity
      tags:
      - sale
    put:
      description: Update an existing sale entity (all required properties must be
        supplied)
      operationId: updateSale
      parameters:
      - description: sale id to update
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/saleUpdate'
        description: Update an existing sale entity
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/saleView'
          description: successful operation
        '400':
          description: Invalid ID
        '404':
          description: sale not found
        '405':
          description: Invalid input
      summary: Update an existing sale entity
      tags:
      - sale
...
security:
- cookieAccessToken: []
  cookieRefreshToken: []
servers:
- url: http://localhost:8001
...